+++ app/templates/items/detail/tabs/location.hbs
@@ -56,9 +56,9 @@
     </div>
     <div class="row content-row box-list">
       {{#reactive-view model=model}}
-        {{#infinite-list height="40vh" loadMore=(action "fetchParentContainers") as |res| }}
-          {{#each res as |pkg|}}
-            {{package-summary-block model=pkg showBoxPalletDetails=true item=model}}
+        {{#infinite-list height="40vh" loadMore=(action "fetchParentContainers") }}
+          {{#each containedPackages as |pkg|}}
+            {{package-summary-block model=pkg showBoxPalletDetails=true item=model updateContainedPackages=(action 'handleChange')}}
           {{/each}}
         {{/infinite-list}}
       {{/reactive-view}}
